Rename directory and header
OneOp test directory: one_op_tests
Header file name: GenModelTest.h
