+++ tests/nnfw_api/src/one_op_tests/AveragePool2D.cc
@@ -14,26 +14,18 @@
  * limitations under the License.
  */
 
-#include "GenModelTests.h"
+#include "GenModelTest.h"
 
-TEST_F(GenModelTest, OneOp_PadV2)
+TEST_F(GenModelTest, OneOp_AvgPool2D)
 {
   CircleGen cgen;
   int in = cgen.addTensor({{1, 2, 2, 1}, circle::TensorType::TensorType_FLOAT32});
-  std::vector<int32_t> padding_data{0, 0, 1, 1, 1, 1, 0, 0};
-  uint32_t padding_buf = cgen.addBuffer(padding_data);
-  int padding = cgen.addTensor({{4, 2}, circle::TensorType::TensorType_INT32, padding_buf});
-  std::vector<float> padding_value_data{3.0};
-  uint32_t padding_value_buf = cgen.addBuffer(padding_value_data);
-  int padding_value =
-      cgen.addTensor({{1}, circle::TensorType::TensorType_FLOAT32, padding_value_buf});
-
-  int out = cgen.addTensor({{1, 4, 4, 1}, circle::TensorType::TensorType_FLOAT32});
-
-  cgen.addOperatorPadV2({{in, padding, padding_value}, {out}});
+  int out = cgen.addTensor({{1, 1, 1, 1}, circle::TensorType::TensorType_FLOAT32});
+  cgen.addOperatorAveragePool2D({{in}, {out}}, circle::Padding_SAME, 2, 2, 2, 2,
+                                circle::ActivationFunctionType_NONE);
   cgen.setInputsAndOutputs({in}, {out});
   _cbuf = cgen.finish();
 
-  _ref_inputs = {{1, 2, 3, 4}};
-  _ref_outputs = {{3, 3, 3, 3, 3, 1, 2, 3, 3, 3, 4, 3, 3, 3, 3, 3}};
+  _ref_inputs = {{1, 3, 2, 4}};
+  _ref_outputs = {{2.5}};
 }
